requirements.txt
Django==1.5.1
defusedxml==0.4.1
distribute==0.6.36
dj-database-url==0.2.1
django-filter==0.6
djangorestframework==2.2.6
gunicorn==0.17.2
psycopg2==2.5
wsgiref==0.1.2
